+++ guillotina/utils/__init__.py
@@ -6,7 +6,8 @@ from .content import get_containers  # noqa
 from .content import get_content_depth  # noqa
 from .content import get_content_path  # noqa
 from .content import get_database  # noqa
-from .content import get_object_by_oid  # noqa
+from .content import get_object_by_uid as get_object_by_oid  # noqa
+from .content import get_object_by_uid  # noqa
 from .content import get_object_url  # noqa
 from .content import get_owners  # noqa
 from .content import iter_databases  # noqa